server/requirements.txt
    openai
    flask
    psycopg2-binary
    pytest
    pytest-flask